seaborn: Drop `numpy` and `matplotlib` as build dependencies.
This seems to fix an issue where NumPy 1.0 is being selected. It seems
like this issue might be related to `conda-build-all` due to this
[line](
https://github.com/SciTools/conda-build-all/blob/459814d56b68648d0a6164558312df4b5fc45e5c/conda_build_all/resolved_distribution.py#L29
). Though it is unclear whether that line actually causes this problem
or if there is some `conda-build` issue that is exposing underneath it
all. In any event, we should keep an eye out for this issue. Going
forward and keep in mind this may in fact be the workaround. We can
probably update `conda-build-all` to pick another default in this case.